requirements.txt
Django==2.2.2
djangorestframework==3.10.2
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
